+++ src/structs/transaction.rs
@@ -1,4 +1,5 @@
 use serde::Deserialize;
+use rust_decimal::Decimal;
 
 use super::DateTime;
 
@@ -28,7 +29,7 @@ pub struct Transaction {
     pub last_updated: DateTime,
     #[serde(rename = "poly")]
     pub transaction_type: TransactionType,
-    pub amount: u32,
+    pub amount: Decimal,
     pub debit_account_field_name: String,
     pub credit_account_field_name: String,
     pub settlement_id: Option<u64>,
@@ -37,13 +38,13 @@ pub struct Transaction {
     pub trade_id: Option<u64>,
     pub group_id: Option<String>,
     pub asset: String,
-    pub debit_pre_balance: Option<u32>,
-    pub debit_post_balance: Option<u32>,
-    pub credit_pre_balance: Option<u32>,
-    pub credit_post_balance: Option<u32>,
+    pub debit_pre_balance: Option<Decimal>,
+    pub debit_post_balance: Option<Decimal>,
+    pub credit_pre_balance: Option<Decimal>,
+    pub credit_post_balance: Option<Decimal>,
     // pub debit_account_mpid: Option<String>,
     pub debit_participant_name: Option<String>,
     // pub credit_account_mpid: Option<String>,
     pub credit_participant_name: Option<String>,
-    pub net_change: i32,
+    pub net_change: Decimal,
 }
